+++ crates/biome_graphql_formatter/tests/spec_test.rs
@@ -36,6 +36,7 @@ pub fn run(spec_input_file: &str, _expected_file: &str, test_directory: &str, _f
                     enabled: Some(true),
                     ..Default::default()
                 }),
+                ..Default::default()
             }),
             ..Default::default()
         },